requirements.txt
# Web Scraping
selenium==4.16.0
beautifulsoup4==4.12.2
# lxml==5.1.0  # Requires C++ compiler on Windows, using html.parser instead
requests==2.31.0

# Browser Drivers
webdriver-manager==4.0.1

# Email
yagmail==0.15.293
python-dotenv==1.0.0

# Database
# sqlite3 is built-in to Python, no installation needed

# Configuration
PyYAML==6.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
fake-useragent==1.4.0

# Logging
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Optional packages (commented out to avoid compilation issues on Windows)
# Uncomment if needed and you have Visual Studio Build Tools installed
# streamlit==1.29.0              # Web dashboard
# python-telegram-bot==20.7      # Telegram notifications
# reportlab==4.0.7               # PDF generation
